Mention slack channels to join in the contributor quickstart
As discussed with Trask in `#team_marketing` slack channel
https://openmined.slack.com/archives/C6G0ZA0HJ/p1502823797000163

Open to suggestions! I was trying to focus on "topics" as opposed to what each
project does, because I think that is redundant with the rest of the
documentation already, and this was more about getting people to
follow and join what is discussed in each channel.